added multi-app install

+++ src/MultiInstall.py
@@ -1,13 +1,10 @@
-from typing import Optional, Callable
-from xml.sax.saxutils import escape
-from gi.repository import Gtk, GObject, Adw, Gdk, Gio, Pango, GLib
-
+import logging
+from gi.repository import Gtk, GObject, Adw, Gdk, Gio, GLib
 
 from .models.AppListElement import InstalledStatus
-from .providers.AppImageProvider import AppImageListElement, AppImageUpdateLogic
+from .providers.AppImageProvider import AppImageListElement
 from .providers.providers_list import appimage_provider
 from .lib.async_utils import _async, idle, debounce
-from .lib.json_config import read_json_config, set_json_config
 from .lib.utils import url_is_valid, get_file_hash, get_application_window
 from .components.AppListBoxItem import AppListBoxItem
 from .components.AppDetailsConflictModal import AppDetailsConflictModal
@@ -16,13 +13,15 @@ from .components.AppDetailsConflictModal import AppDetailsConflictModal
 class MultiInstall(Gtk.ScrolledWindow):
     __gsignals__ = {
         "show-details": (GObject.SIGNAL_RUN_FIRST, GObject.TYPE_NONE, (object, )),
+        "go-back": (GObject.SIGNAL_RUN_FIRST, GObject.TYPE_NONE, (bool, )),
     }
 
     def __init__(self) -> None:
         super().__init__()
         self.ACTION_ROW_ICON_SIZE = 45
 
-        self.app_list: list[AppListBoxItem] = []
+        self.app_list: list[AppImageListElement] = []
+        self.app_list_box_items: list[AppListBoxItem] = []
         self.app_list_box = Gtk.ListBox(css_classes=['boxed-list'])
 
         self.install_all_btn = Gtk.Button(
@@ -32,11 +31,11 @@ class MultiInstall(Gtk.ScrolledWindow):
 
         self.main_box = Gtk.Box(
             orientation=Gtk.Orientation.VERTICAL, 
-            margin_top=10, 
+            margin_top=30, 
             margin_bottom=10, 
             margin_start=20, 
             margin_end=20,
-            spacing=10
+            spacing=20
         )
 
         self.main_box.append(self.install_all_btn)
@@ -49,14 +48,26 @@ class MultiInstall(Gtk.ScrolledWindow):
             halign=Gtk.Align.CENTER,
             child=Adw.ButtonContent(
                 label=_('Some apps are already installed'),
-                icon_name='software-update-available-symbolic'
+                icon_name='info-symbolic'
             )
         )
 
         self.main_box.append(self.already_installed_warn)
 
-        clamp= Adw.Clamp(child=self.main_box)
-        self.set_child(clamp)
+        clamp = Adw.Clamp(child=self.main_box)
+        overlay = Gtk.Overlay(
+            child=clamp,
+        )
+
+
+        self.progress_bar = Gtk.ProgressBar(
+            css_classes=['osd', 'horizontal', 'top'], fraction=1
+        )
+
+        overlay.add_overlay(self.progress_bar)
+        overlay.set_clip_overlay(self.progress_bar, True)
+
+        self.set_child(overlay)
 
     @idle
     def create_app_row_complete_load(self, el: AppImageListElement, icon: Gtk.Image):
@@ -68,8 +79,26 @@ class MultiInstall(Gtk.ScrolledWindow):
 
         row.details_btn.connect('clicked', self.on_details_btn_clicked, el)
 
-        self.app_list.append(row)
         self.app_list_box.append(row)
+        self.app_list_box_items.append(row)
+
+        fraction = len(self.app_list_box_items) / len(self.app_list)
+        self.progress_bar.set_fraction(fraction)
+
+        if fraction == 1:
+            self.progress_bar.set_visible(False)
+
+            not_installed_count = self.count_not_installed()
+
+            self.install_all_btn.set_sensitive(not_installed_count > 0)
+
+    def count_not_installed(self):
+        not_installed_count = 0
+        for el in self.app_list:
+            if el.installed_status is InstalledStatus.NOT_INSTALLED:
+                not_installed_count += 1
+
+        return not_installed_count
 
     @_async
     def create_app_row(self, el: AppImageListElement):
@@ -80,9 +109,13 @@ class MultiInstall(Gtk.ScrolledWindow):
 
     def show_confirmation_dialog(self):
         dialog = Adw.MessageDialog(
-            parent=get_application_window(),
+            transient_for=get_application_window(),
             heading=_('Do you really want to move all the apps to the menu?')
         )
+
+        dialog.add_response('cancel', _('Cancel'))
+        dialog.add_response('confirm', _('Proceed'))
+        dialog.set_response_appearance('confirm', Adw.ResponseAppearance.SUGGESTED)
 
         body = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
         checkbox = Gtk.CheckButton.new_with_label(_('I have verified the source of the apps'))
@@ -90,29 +123,68 @@ class MultiInstall(Gtk.ScrolledWindow):
 
         body.append(checkbox)
         dialog.set_extra_child(body)
+        dialog.connect('response', self.on_dialog_response, checkbox)
 
         dialog.present()
+
+    def on_dialog_response(self, dialog: Adw.MessageDialog, response: str, checkbox: Gtk.CheckButton):
+        if not checkbox.get_active():
+            return
+        
+        if response != 'confirm':
+            return
+        
+        self.install_all_btn.set_sensitive(False)
+        for el in self.app_list:
+            if el.installed_status is InstalledStatus.INSTALLED:
+                continue
+
+            el.set_trusted()
+            appimage_provider.install_file(el)
+
+        self.emit('go-back', True)
 
     def on_details_btn_clicked(self, widget: Gtk.Button, el: AppImageListElement):
         self.emit('show-details', el)
 
-    def set_from_local_files(self, files: list[Gio.File]):
-        self.app_list_box.remove_all()
-        self.app_list = []
-
-        not_installed_count = 0
-
+    @_async
+    def create_list_elements(self, files):
         for f in files:
-            el = appimage_provider.create_list_element_from_file(f)
+            try:
+                el = appimage_provider.create_list_element_from_file(f)
+                self.app_list.append(el)
+
+            except Exception as e:
+                logging.error(e)
+
+        self.progress_bar.set_visible(True)
+        for el in self.app_list:
             self.create_app_row(el)
 
-            if el.installed_status is InstalledStatus.NOT_INSTALLED:
-                not_installed_count += 1
+        not_installed_count = self.count_not_installed()
+        show_warn = not_installed_count != len(self.app_list)
+        
+        GLib.idle_add(lambda: 
+            self.already_installed_warn.set_visible(show_warn))
 
-        self.install_all_btn.set_sensitive(not_installed_count > 0)
-        self.already_installed_warn.set_visible(not_installed_count == 0)
+    def set_from_local_files(self, files: list[Gio.File]):
+        if self.progress_bar.get_fraction() not in [0, 1]:
+            return True
 
-        return True
+        self.progress_bar.pulse()
+        self.app_list_box.remove_all()
+        self.app_list = []
+        self.app_list_box_items = []
+
+        self.install_all_btn.set_sensitive(False)
+
+        installable = 0
+        for f in files:
+            if appimage_provider.can_install_file(f):
+                installable += 1
+
+        self.create_list_elements(files)
+        return installable > 0
 
     def on_install_all_clicked(self, widget: Gtk.Button):
         body = Gtk.BaselinePosition
